+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Release v0.4.0 (Under Development)
 
-* Added new error handling and recovery with composable labeled failures, recovery rules and error handlers.
+* Added new error handling and recovery with labeled failures, recovery rules and error handlers.
 * Refactored the encoder expression interfaces into CRTP base classes to allow for common functionality to be shared by all concrete encoder expressions, such as failure labeling and recovery rules.
+* Added new `accept` & `accept_then_cut` operators that work in conjuction with cut operator. Tested and fixed issues with cut and accept.
+* Added `on_reset` and `on_drain hooks` to the `lug::environment` class.
 * Implemented new 64-bit instruction encoding scheme that is simpler and more efficient. String data is no longer embedded in the instruction stream, ensuring constant instruction length of 8 bytes.
 * Merged `lug::program_encoder` and `lug::rule_encoder` into the `lug::encoder` base class after overhauling how choice/jump offsets are calculated, significantly reducing binary size bloat.
 * Extracted common base class of `lug::basic_parser` into `lug::parser_base` to reduce template bloat.